Rename files for julia and mandelbrot to zsquare seed and param respectively
