+++ include/fractool/ftensp/zsquare-seed.hpp
