+++ src/ftensp/CMakeLists.txt
@@ -1,5 +1,5 @@
 add_library(ftensp
 	algorithm.cpp
-	generate_julia.cpp
-	generate_mandelbrot.cpp)
+	zsquare-seed.cpp
+	zsquare-param.cpp)
 target_link_libraries(ftensp PUBLIC ftcore)
